🎨 Small refactor to standardize my config style
## Changes Description

I decided to standardize the way I configure plugins in this config.

Basically, if a plugin requires me to have the `config` attribute
defined, I always set it **before** the `opts` one.

I also renamed some variables to a better name, and made sure I only
require libraries from the plugins being installed as part of their
configuration functions, and not as a global variable.